Refactor event_hook
