+++ src/ui/event_hook.c
@@ -1,0 +1,33 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   event_hook.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/10/13 23:34:34 by reasuke           #+#    #+#             */
+/*   Updated: 2024/10/13 23:35:54 by reasuke          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include <signal.h>
+#include <stdio.h>
+#include <unistd.h>
+
+#include "libft.h"
+#include "readline/readline.h"
+#include "ui.h"
+
+int	handle_sigint_hook(void)
+{
+	if (g_signum == SIGINT)
+	{
+		g_signum = 0;
+		ft_putchar_fd('\n', STDOUT_FILENO);
+		rl_replace_line("", 0);
+		rl_on_new_line();
+		rl_redisplay();
+	}
+	return (0);
+}
+
