+++ src/ui/signal_handler.c
@@ -6,7 +6,7 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/23 20:03:51 by reasuke           #+#    #+#             */
-/*   Updated: 2024/10/13 23:30:38 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/10/13 23:34:40 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,19 +21,6 @@
 #include "utils.h"
 
 volatile sig_atomic_t	g_signum = 0;
-
-int	handle_sigint_hook(void)
-{
-	if (g_signum == SIGINT)
-	{
-		g_signum = 0;
-		ft_putchar_fd('\n', STDOUT_FILENO);
-		rl_replace_line("", 0);
-		rl_on_new_line();
-		rl_redisplay();
-	}
-	return (0);
-}
 
 static void	set_signum(int sig)
 {
